Got working on os x
Had problems with regex in bash on my mac. First upgraded to bash 4
in ports but still had to modify the regex commands. Reverted back
to new syntax (from expr) because didn't know about $BASH_REMATCH
initially. This'll hopefully work in linux too. Seem to have to
store the regex in a variable, inline doesn't work

Need to check what alt-r is doing, some commands being ignored
which are not duplicates